Refactor - Moving for a better structure
