+++ src/app/app-routing.module.ts
@@ -1,21 +1,23 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
-import { ListProductsModule } from './modules/list-products/list-products.module';
+import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
+import { ListProductsModule } from './products/list-products.module';
 
 export const routes: Routes = [
   {
     path: 'a',
-    loadChildren: () => import('./modules/a/a.module').then((m) => m.AModule)
+    loadChildren: () =>
+      import('./routing/modules/a/a.module').then((m) => m.AModule)
   },
   {
     path: 'b',
-    loadChildren: () => import('./modules/b/b.module').then((m) => m.BModule)
+    loadChildren: () =>
+      import('./routing/modules/b/b.module').then((m) => m.BModule)
   },
   {
     path: 'products',
     loadChildren: () =>
-      import('./modules/list-products/list-products-routing.module').then(
+      import('./products/list-products-routing.module').then(
         (m) => m.ListProductsRoutingModule
       )
   },
