+++ src/app/products/list-products/list-products.component.scss
