+++ src/app/products/paginator/paginator.component.html
