+++ src/app/products/paginator/paginator.component.scss
