+++ src/app/products/paginator/paginator.component.ts
