+++ src/app/products/prices/prices.component.html
