+++ src/app/products/prices/prices.component.scss
