+++ src/app/products/prices/prices.component.ts
