+++ src/app/products/product-detail/product-detail.component.scss
