+++ src/app/routing/components/a1/a1.component.ts
