+++ src/app/routing/components/b2/b2.component.spec.ts
