+++ src/app/routing/modules/a/a-routing.module.ts
@@ -23,4 +23,4 @@ export const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class ARoutingModule { }
+export class ARoutingModule {}
