+++ src/app/routing/modules/b/b-routing.module.ts
@@ -23,4 +23,4 @@ export const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class BRoutingModule { }
+export class BRoutingModule {}
