+++ src/app/routing/modules/b/b.module.ts
@@ -7,17 +7,8 @@ import { B2Component } from '../../components/b2/b2.component';
 import { B3Component } from '../../components/b3/b3.component';
 import { ZModule } from '../z/z.module';
 
-
 @NgModule({
-  declarations: [
-    B1Component,
-    B2Component,
-    B3Component
-  ],
-  imports: [
-    CommonModule,
-    BRoutingModule,
-    ZModule
-  ]
+  declarations: [B1Component, B2Component, B3Component],
+  imports: [CommonModule, BRoutingModule, ZModule]
 })
-export class BModule { }
+export class BModule {}
